+++ src/Serializer/JsonSerializerInterface.php
@@ -20,24 +20,20 @@
 
 declare(strict_types=1);
 
-namespace OAT\Library\Lti1p3Ags\Model\Result;
+namespace OAT\Library\Lti1p3Ags\Serializer;
 
 use JsonSerializable;
+use RuntimeException;
 
-/**
- * @see https://www.imsglobal.org/spec/lti-ags/v2p0#result-service
- */
-interface ResultContainerInterface extends JsonSerializable
+interface JsonSerializerInterface
 {
-    public const REL_NEXT = 'next';
+    /**
+     * @throws RuntimeException
+     */
+    public function serialize(JsonSerializable $object): string;
 
-    public function getResults(): ResultCollectionInterface;
-
-    public function getRelationLink(): ?string;
-
-    public function setRelationLink(?string $relationLink): ResultContainerInterface;
-
-    public function getRelationLinkUrl(): ?string;
-
-    public function hasNext(): bool;
+    /**
+     * @throws RuntimeException
+     */
+    public function deserialize(string $json): array;
 }